% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.R
\name{AIC.cvRemix}
\alias{AIC.cvRemix}
\title{AIC for cvRemix object}
\usage{
\method{AIC}{cvRemix}(x)
}
\arguments{
\item{x}{output of \code{\link{cv.remix}}.}
}
\value{
AIC across the grid.
}
\description{
Computes akaike information criterion from the output of \code{\link{cv.remix}}as
\deqn{AIC = -2\mathcal{LL}_{y}(\hat\theta,\hat\alpha)+2P}
where \eqn{P} is the total number of parameters estimated and \eqn{\mathcal{LL}_{y}(\hat\theta,\hat\alpha)} the log-likelihood of the model.
}
\examples{
\dontrun{
project <- getMLXdir()

ObsModel.transfo = list(S=list(AB=log10),
                        linkS="yAB",
                        R=rep(list(S=function(x){x}),5),
                        linkR = paste0("yG",1:5))

alpha=list(alpha0=NULL,
           alpha1=setNames(paste0("alpha_1",1:5),paste0("yG",1:5)))

y = c(S=5,AB=1000)

res = cv.remix(project = project,
               dynFUN = dynFUN_demo,
               y = y,
               ObsModel.transfo = ObsModel.transfo,
               alpha = alpha,
               selfInit = TRUE,
               eps1=10**(-2),
               ncores=8,
               nlambda=8,
               eps2=1)

AIC(res)
}
}
\references{
Akaike, H. 1998. Information theory and an extension of the maximum likelihood principle, Selected papers of hirotugu akaike, 199–213. New York: Springer.
}
