% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.R
\name{BIC.cvRemix}
\alias{BIC.cvRemix}
\title{BIC for cvRemix object}
\usage{
\method{BIC}{cvRemix}(x)
}
\arguments{
\item{x}{output of \code{\link{cv.remix}}.}
}
\value{
BIC across the grid.
}
\description{
Computes bayesian information criterion from the output of \code{\link{cv.rremix}} as
\deqn{BIC = -2\mathcal{LL}_{y}(\hat\theta,\hat\alpha)+\log(N)P}
where \eqn{P} is the total number of parameters estimated, \eqn{N} the number of subject and \eqn{\mathcal{LL}_{y}(\hat\theta,\hat\alpha)} the log-likelihood of the model.
}
\examples{
\dontrun{
project <- getMLXdir()

ObsModel.transfo = list(S=list(AB=log10),
                        linkS="yAB",
                        R=rep(list(S=function(x){x}),5),
                        linkR = paste0("yG",1:5))

alpha=list(alpha0=NULL,
           alpha1=setNames(paste0("alpha_1",1:5),paste0("yG",1:5)))

y = c(S=5,AB=1000)

res = cv.remix(project = project,
               dynFUN = dynFUN_demo,
               y = y,
               ObsModel.transfo = ObsModel.transfo,
               alpha = alpha,
               selfInit = TRUE,
               eps1=10**(-2),
               ncores=8,
               nlambda=8,
               eps2=1)

BIC(res)
}
}
\references{
Schwarz, G. 1978. Estimating the dimension of a model. The annals of statistics 6 (2): 461–464
}
