% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irene2inj.R
\name{Covid}
\alias{Covid}
\title{Title}
\usage{
Covid(
  t,
  S0,
  A0,
  t_inj,
  fM1,
  fM2,
  theta1,
  theta_2,
  delta_S,
  delta_V = 2.7,
  delta_Ab = 0.03
)
}
\arguments{
\item{t}{time}

\item{S0}{s0}

\item{A0}{a}

\item{t_inj}{a}

\item{fM1}{a}

\item{fM2}{a}

\item{theta1}{a}

\item{theta_2}{a}

\item{delta_S}{a}

\item{delta_V}{a}

\item{delta_Ab}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a
}
