% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.R
\name{eBIC.cvRemix}
\alias{eBIC.cvRemix}
\title{eBIC for cvRemix object}
\usage{
\method{eBIC}{cvRemix}(x, gamma = 1)
}
\arguments{
\item{x}{output of \code{\link{remix}}.}
}
\value{
eBIC accross the grid.
}
\description{
Computes extended bayesian information criterion from the output of \code{\link{cv.remix}}.
}
\examples{
\dontrun{
project <- getMLXdir()

ObsModel.transfo = list(S=list(Ab=log10),
                        linkS="yAB",
                        R=rep(list(S=function(x){x}),5),
                        linkR = paste0("yG",1:5))

alpha=list(alpha0=NULL,
           alpha1=setNames(paste0("alpha_1",1:5),paste0("yG",1:5)))

y = c(S=5,Ab=1000)

res = cv.remix(project = project,
               dynFUN = dynFUN_demo,
               y = y,
               ObsModel.transfo = ObsModel.transfo,
               alpha = alpha,
               selfInit = TRUE,
               eps1=10**(-2),
               ncores=8,
               nlambda=8,
               eps2=1)

eBIC(res)
}
}
\references{
Chen, J., & Chen, Z. (2008). Extended Bayesian Information Criteria for Model Selection with Large Model Spaces. Biometrika, 95(3), 759–771. http://www.jstor.org/stable/20441500
}
