% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoFunction.R
\name{getMLXdir}
\alias{getMLXdir}
\title{Get monolix demo project path}
\usage{
getMLXdir()
}
\value{
path to the monolix demo from REMix package.
}
\description{
Get monolix demo project path
}
\examples{
print(getMLXdir())
}
\seealso{
\code{\link{dynFUN_demo}}.
}
