% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_int.R
\name{gh.int.ind}
\alias{gh.int.ind}
\title{Title}
\usage{
gh.int.ind(
  Omega,
  theta,
  alpha1,
  dynFun,
  identifier.SR,
  y,
  covariates,
  ParModel.transfo,
  ParModel.transfo.inv,
  Sobs,
  Robs,
  Serr,
  Rerr,
  ObsModel.transfo,
  n = 3,
  prune = NULL
)
}
\arguments{
\item{Omega}{a}

\item{theta}{a}

\item{alpha1}{a}

\item{dynFun}{a}

\item{y}{a}

\item{covariates}{a}

\item{ParModel.transfo}{a}

\item{ParModel.transfo.inv}{aa}

\item{Sobs}{a}

\item{Robs}{a}

\item{Serr}{a}

\item{Rerr}{a}

\item{ObsModel.transfo}{a}

\item{n}{a}

\item{prune}{a}

\item{ObsModel.transfo.inv}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a
}
