% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indparm.R
\name{indParm}
\alias{indParm}
\title{Generate individual parameters}
\usage{
indParm(theta, covariates, eta_i, transfo, transfo.inv)
}
\arguments{
\item{theta}{list with at least \code{phi_pop}, \code{psi_pop}, \code{gamma}, \code{beta} (named ; corresponding to the model parameter \eqn{\phi_{pop}}, \eqn{\psi_{pop}}, \eqn{\gamma}, \eqn{\beta} ) : \itemize{\item \code{phi_pop} named vector of population parameters without r.e ;\item \code{psi_pop} named vector of population parameters with r.e ;\item \code{gamma} named list of vector of covariates effects for  \code{phi_pop} parameters, if NULL no covariates effect on parameters. ;\item \code{beta} named list of vector of covariates effects for each \code{psi_pop}, if NULL no covariates effect on parameters.}}

\item{covariates}{line data.frame of individual covariates ;}

\item{eta_i}{named vector of random effect for each \code{psi} parameter ;}

\item{transfo}{named list of transformation functions \eqn{(h_l)_{l\leq m}} and \eqn{(s_k)_{k\leq K}} for the individual parameter model (names must be consistent with \code{phi_pop} and \code{psi_pop}, missing entries are set by default to the identity function).}

\item{transfo.inv}{amed list of inverse transformation functions for the individual parameter model (names must be consistent with\code{phi_pop} and\code{psi_pop}).}
}
\value{
a list with \code{phi_i} and \code{psi_i} parameters.
}
\description{
Generate the individual parameters of indivual whose covariates are \code{covariates} and random effects \code{eta_i}.
}
\details{
The models used for the parameters are :
\deqn{ h_l(\psi_{li}) = h_l(\psi_{lpop})+X_i\beta_l + \eta_{li} }
with \eqn{h_l} the transformation, \eqn{\beta_l} the vector of covariates effect and  with \eqn{\eta_i} the random effects associated \eqn{\psi_l} parameter ;
\deqn{ g_k(\phi_{ki}) = g_k(\phi_{kpop})+X_i \gamma_l }
with \eqn{g_k} the transformation and  \eqn{\gamma_k} the vector of covariates effect associated \eqn{\phi_k} parameter.
}
\examples{
phi_pop = c(delta_S = 0.231, delta_L = 0.000316)
psi_pop = c(delta_Ab = 0.025,phi_S = 3057, phi_L = 16.6)
gamma = NULL
covariates = data.frame(cAGE = runif(1,-15,15), G1 = rnorm(1), G2 = rnorm(1))
beta = list(delta_Ab=c(0,1.2,0),phi_S = c(0.93,0,0),phi_L=c(0,0,0.8))

theta=list(phi_pop = phi_pop,psi_pop = psi_pop,gamma = gamma, beta = beta)
eta_i = c(delta_Ab = rnorm(1,0,0.3),phi_S=rnorm(1,0,0.92),phi_L=rnorm(1,0,0.85))
transfo = list(delta_Ab=log,phi_S=log,phi_L=log)
transfo.inv = list(delta_Ab = exp,phi_S=exp,phi_L=exp)

indParm(theta,covariates,eta_i,transfo,transfo.inv)
}
\seealso{
\code{\link{model.clairon}}, \code{\link{model.pasin}}.
}
