% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoLLPen-package.R
\docType{package}
\name{LassoLLPen-package}
\alias{LassoLLPen}
\alias{LassoLLPen-package}
\title{LassoLLPen: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
