% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pasin.R
\name{Pasin}
\alias{Pasin}
\title{Antibodies concentration}
\usage{
Pasin(t, A0, phi_S, phi_L, delta_Ab, delta_S = 0.23, delta_L = 0.000316)
}
\arguments{
\item{t}{time}

\item{A0}{antibody at baseline}

\item{phi_S}{phi_S}

\item{phi_L}{phi_L}

\item{delta_Ab}{delta_Ab}

\item{delta_S}{delta_S}

\item{delta_L}{delta_L}
}
\value{
éa
}
\description{
Antibodies concentration
}
\examples{
a
}
