% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cvRemix}
\alias{plot.cvRemix}
\title{Plot of cv.remix object}
\usage{
\method{plot}{cvRemix}(x, criterion = BICc, trueValue = NULL)
}
\arguments{
\item{x}{output of \code{\link{cv.remix}}.}
}
\value{
A plot.
}
\description{
Calibration and Informative Criterion plot for cvRemix object.
}
\seealso{
\code{\link{cv.remix}}
}
