% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Log-likelihood convergence.}
\usage{
plotConvergence(fit)
}
\arguments{
\item{fit}{fit object of class remix, from \code{\link{Remix}} or a certain build from \code{\link{cv.Remix}} output.}
}
\value{
Log-Likelihood values throughout the algorithm iteration.
}
\description{
Log-likelihood convergence.
}
\examples{
## [ TO DO ]
}
\seealso{
\code{\link{Remix}}, \code{\link{cv.Remix}}.
}
