% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotIC}
\alias{plotIC}
\title{IC plot}
\usage{
plotIC(fit, criterion = BICc, dismin = TRUE)
}
\arguments{
\item{fit}{fit object of class cvRemix, from \code{\link{cv.remix}};}

\item{criterion}{which criterion among 'BICc', 'BIC', 'AIC' or 'eBIC' to take into account (default: BICc);}

\item{dismin}{logical ; if minimizers of information criterion should be display.}
}
\value{
IC trhoughout the lambda.grid.
}
\description{
IC plot
}
\examples{
\dontrun{
project <- getMLXdir()

ObsModel.transfo = list(S=list(AB=log10),
                        linkS="yAB",
                        R=rep(list(S=function(x){x}),5),
                        linkR = paste0("yG",1:5))

alpha=list(alpha0=NULL,
           alpha1=setNames(paste0("alpha_1",1:5),paste0("yG",1:5)))

y = c(S=5,AB=1000)

res = cv.remix(project = project,
               dynFUN = dynFUN_demo,
               y = y,
               ObsModel.transfo = ObsModel.transfo,
               alpha = alpha,
               selfInit = TRUE,
               eps1=10**(-2),
               ncores=8,
               nlambda=8,
               eps2=1)

plotCalibration(res)

plotIC(res)
}
}
\seealso{
\code{\link{remix}}, \code{\link{cv.remix}}.
}
