% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSAEM}
\alias{plotSAEM}
\title{Display the value of parameters at each iteration}
\usage{
plotSAEM(fit, paramToPlot = "all", trueValue = NULL)
}
\arguments{
\item{fit}{object of class remix, from \code{\link{Remix}} or a certain build from \code{\link{cv.Remix}} output.}

\item{paramToPlot}{Population parameters to plot (which have been estimated by SAEM) ;}

\item{trueValue}{(for simulation purpose) vector named of true values ;}
}
\value{
For each parameters, the values at the end of each iteration of remix algorithm is drawn. Moreover, the SAEM steps of each iteration are displayed.
}
\description{
Display the value of parameters at each iteration
}
\examples{
# # [ TO DO ]
}
\seealso{
\code{\link{Remix}}, \code{\link{cv.Remix}}.
}
