% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clairon.R
\name{Clairon}
\alias{Clairon}
\title{Generate trajectory of antibody response to SARS-Cov vaccine.}
\usage{
Clairon(t, y, parms)
}
\arguments{
\item{t}{vector of time ;}

\item{y}{initial condition, named vector of form c(S=S0,A=A0) ;}

\item{parms}{named vector of model parameter ; should contain "t_0","t_inj","fM2","theta","delta_S","delta_Ab","delta_V".}
}
\value{
Matrix of time and observation of antibody titer Ab.
}
\description{
The model correspond to the antibody response to two injection of vaccine at time `t_0`and `t_inj`.Model is defined as
\deqn{\displaystyle\left\{\begin{matrix} \frac{d}{dt} S(t) &=& f_{\overline M_k} e^{-\delta_V(t-t_k)}-\delta_S S(t) \\ \frac{d}{dt} Ab(t) &=& \theta S(t) - \delta_{Ab} Ab(t)\end{matrix}\right.  }
on each interval \eqn{I_1=[t_0;t_{inj}[ } and \eqn{I_2=[t_{inj};+\infty[}. For each interval \eqn{I_k}, we have \eqn{t_k} corresponding to the last injection date of vaccine, such that \eqn{t_1=t_0} and \eqn{t_2=t_{inj}}. By definition, \eqn{f_{\overline M_1}=1} (Clairon and al., 2023).
}
\examples{
y = c(S=1,A=0)

parms = c(fM2 = 4.5,
          theta = 18.7,
          delta_S = 0.01,
          delta_Ab = 0.23,
          delta_V = 2.7)

t = seq(0,35,1)

res <- Clairon(t,y,parms)

plot(res)

}
\references{
Quentin Clairon, Mélanie Prague, Delphine Planas, Timothée Bruel, Laurent Hocqueloux, et al.. Modeling the evolution of the neutralizing antibody response against SARS-CoV-2 variants after several administrations of Bnt162b2. 2023. hal-03946556
}
