% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.remix.R
\name{cv.Remix}
\alias{cv.Remix}
\title{REMix algorithm}
\usage{
cv.Remix(
  project = NULL,
  final.project = NULL,
  dynFUN,
  y,
  ObsModel.transfo,
  alpha,
  lambda.grid = NULL,
  nlambda = 50,
  eps1 = 10^(-1),
  eps2 = 10^(-1),
  selfInit = FALSE,
  pop.set1 = NULL,
  pop.set2 = NULL,
  prune = NULL,
  n = NULL,
  ncores = NULL,
  print = TRUE,
  digits = 3,
  trueValue = NULL
)
}
\arguments{
\item{project}{the initial Monolix project;}

\item{final.project}{the final Monolix project (by default adds "_upd" to the original project), every useful log is saved in the remix folder directly in the initial project directory.}

\item{y}{Initial condition of the model, conform to what is asked in dynFun ;}

\item{ObsModel.transfo}{list of 2 list of P,K transformation (need to include identity transformation), named with `S` and `R` :

  - ObsModel.transfo$S correspond to the transformation used for direct observation model. For each \eqn{Y_p=h_p(S_p)} the order (as in Sobs) must be respected and the name indicated which dynamic from dynFun is observed through this variables \eqn{Y_p};

  - ObsModel.transfo$R correspond to the transformation used for the latent process, as it is now, we only have one latent dynamic so necessarily \eqn{s_k} is applied to `R` but for each \eqn{Y_k} observed, transformation could be different so need to precise as many as in Robs ; the name need be set to precise the dynamic from dynFun to identify the output.}

\item{alpha}{list of named vector "alpha0", "alpha1" (in good order), alpha1 mandatory even if 1}

\item{lambda.grid}{grid of penalisation parameters for lasso regularization;}

\item{nlambda}{if lambda.grid is null, number of lambda parameter to use for the lambda.grid;}

\item{eps1}{convergence limit for parameters distance at each iteraion ;}

\item{eps2}{convergence limit for penalized log-likelihood distance at each iteration ;}

\item{pop.set1}{Population parameters setting for initialisation ;}

\item{pop.set2}{Population parameters setting for algorithm iterations ;}

\item{prune}{(default NULL) percentage in [0;1] for prunning.}

\item{n}{(default floor(100**(1/length(theta$psi_pop))) number of points for gaussian quadrature ;}

\item{ncores}{number of cores for parallelization (default NULL).}

\item{print}{if TRUE, log are printed in console. Logs are allways saved in a summary file in the remix folder created for the job;}

\item{digits}{digits to print, (default 3) ;}

\item{trueValue}{(FOR SIMULATION, if provided, the error is compute at each iteration. )}

\item{dynFun}{Dynamic function ;}
}
\value{
list fo outputs of final project and through the iteration for every lambda on lambda.grid.
}
\description{
Regularization and Estimation in MIXed effects model, over a regularization path.
}
\examples{
[ TO DO ]
}
