% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasin.R
\name{Pasin}
\alias{Pasin}
\title{Generate trajectory of the Humoral Immune Response to a Prime-Boost Ebola Vaccine}
\usage{
Pasin(t, y, parms)
}
\arguments{
\item{t}{vector of time ;}

\item{y}{initial condition, named vector of form c(A=A0,S=S0,L=L0) ;}

\item{parms}{named vector of model parameter ; should contain ""theta_S","theta_L","delta_Ab","delta_S","delta_L".}
}
\value{
Matrix of time and observation of antibody titer Ab, and ASCs S and L.
}
\description{
The model correspond to the dynamics of the humoral response, from 7 days after the boost immunization. Model is defined as
\deqn{ \left\{\begin{matrix}\frac{d}{dt} Ab(t) &=& \theta_S S(t) + \theta_L L(t) - \delta_{Ab} Ab(t)  \\ \frac{d}{dt}S(t) &=& -\delta_S S(t) \\ \frac{d}{dt} L(t) &=& -\delta_L L(t)\end{matrix}\right. }
}
\examples{
y = c(A=0,S=5,L=5)
parms = c(theta_S = 611,
       theta_L = 3.5,
              delta_Ab = 0.025,
              delta_S = 0.231,
              delta_L = 0.000152)

t = seq(0,100,5)
res <- Pasin(t,y,parms)
plot(res)

}
\references{
Pasin C, Balelli I, Van Effelterre T, Bockstal V, Solforosi L, Prague M, Douoguih M, Thiébaut R, for the EBOVAC1 Consortium. 2019. Dynamics of the humoral immune response to a prime-boost Ebola vaccine: quantification and sources of variation. J Virol 93:e00579-19. https://doi.org/10.1128/JVI.00579-19.
}
\seealso{
[indParm()]
}
