% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PK.R
\name{PK}
\alias{PK}
\title{Individual PK trajectory}
\usage{
PK(t, C0, Cl = NULL, V = NULL, ka = Cl/V)
}
\arguments{
\item{t}{vector of observation times.}

\item{C0}{Initial concentration $C(t=0)$.}

\item{Cl}{Cleareance of subject.}

\item{V}{Compartment volume of subject.}

\item{ka}{ka value}
}
\value{
The concentration $C(t)$ of subject defined by individual parameters $Cl$, $V$ and $C_0$.
}
\description{
The administration is via a bolus.
The PK model has one compartment (volume $V$) and a linear elimination (clearance $Cl$).
}
\examples{
a
}
