% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMLX.R
\name{readMLX}
\alias{readMLX}
\title{Gather information from a monolix project}
\usage{
readMLX(project = NULL, ObsModel.transfo, alpha)
}
\arguments{
\item{project}{directory of the monolix project (in .mlxtran), if NULL get current project load, default to NULL}

\item{ObsModel.transfo}{list of 2 list of P,K transformation (need to include identity transformation), named with `S` and `R` :

  - ObsModel.transfo$S correspond to the transformation used for direct observation model. For each \eqn{Y_p=h_p(S_p)}, the name indicate which dynamic from monolix is observed through this variables \eqn{Y_p};

  - ObsModel.transfo$R correspond to the transformation used for the latent process, as it is now, we only have one latent dynamic so necessarily \eqn{s_k} is applied to `R` but for each \eqn{Y_k} observed, transformation could be different so need to precise as many as in the monolix project created ;}

\item{alpha}{list of named vector "alpha0", "alpha1" (in good order), alpha1 mandatory even if 1}
}
\value{
list with parameters, transformations, observations from the monolix project, with the form needed for the implemented algorithm
}
\description{
Retrieve all information necessary for the algorithm that can be search in the created monolix project.
}
\examples{
# TODO

}
